Added support for string and string typecast.   -> Added new test cases related to string and string type cast.   -> modified some code in constant load class.

+++ BIR2llvmIR/BIRReader.h
@@ -142,10 +142,10 @@ public:
   ~FloatCpInfo();
 
 private:
-  float value;
+  double value;
 
 public:
-  float getValue() const { return value; }
+  double getValue() const { return value; }
 };
 
 class ByteCpInfo : public ConstantPoolEntry {
